Access error messages from nested responses
Some responses from the filestore will return a nested response which
can contain fields whose values are used to present the correct
validation error message to a user.

For example, when there is a validiation error on a file upload as a
result of the file being an unsupported type, the response from the
filestore will be parsed into an error object that has further response
body field containing a 'name' field. For unsupported file types the
value of this will be 'accept' ( ¯\_(ツ)_/¯ ). This value maps to custom
error messages which can be created by the form owners to be shown to
their end users.

Here we attempt to retreive the necessary values, but fall back to
defaults if we are unable to do so.